# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HomeBites - Home food and groceries marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  message: string;
}

export type CategoryKind = (typeof CategoryKind)[keyof typeof CategoryKind];

export const CategoryKind = {
  home_food: "home_food",
  grocery: "grocery",
} as const;

export interface Category {
  id: number;
  name: string;
  slug: string;
  kind: CategoryKind;
  imageUrl: string;
}

export interface Chef {
  id: number;
  name: string;
  tagline: string;
  cuisine: string;
  rating: number;
  etaMinutes: number;
  deliveryFee: number;
  imageUrl: string;
  location: string;
  priceForTwo: number;
  isVeg: boolean;
  featured?: boolean;
}

export type DishSpiceLevel =
  (typeof DishSpiceLevel)[keyof typeof DishSpiceLevel];

export const DishSpiceLevel = {
  mild: "mild",
  medium: "medium",
  spicy: "spicy",
} as const;

export interface Dish {
  id: number;
  chefId: number;
  chefName: string;
  name: string;
  description: string;
  price: number;
  imageUrl: string;
  isVeg: boolean;
  rating: number;
  spiceLevel: DishSpiceLevel;
}

export interface ChefDetail {
  chef: Chef;
  dishes: Dish[];
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  mrp: number;
  unit: string;
  imageUrl: string;
  categoryId: number;
  categoryName: string;
  inStock: boolean;
  essential?: boolean;
}

export interface Offer {
  id: number;
  title: string;
  subtitle: string;
  code: string;
  accentColor: string;
}

export interface SearchResults {
  chefs: Chef[];
  dishes: Dish[];
  products: Product[];
}

export type OrderItemKind = (typeof OrderItemKind)[keyof typeof OrderItemKind];

export const OrderItemKind = {
  dish: "dish",
  product: "product",
} as const;

export interface OrderItem {
  kind: OrderItemKind;
  refId: number;
  name: string;
  imageUrl: string;
  unitPrice: number;
  quantity: number;
}

export interface Address {
  fullName: string;
  phone: string;
  line1: string;
  line2?: string;
  city: string;
  pincode: string;
  landmark?: string;
}

export interface MeResponse {
  id?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
  imageUrl?: string;
  isAdmin: boolean;
}

export type AdminStatsOrdersByStatus = { [key: string]: number };

export type AdminStatsRevenueByDayItem = {
  day: string;
  revenue: number;
  orders: number;
};

export interface AdminStats {
  totalOrders: number;
  totalRevenue: number;
  totalProducts: number;
  totalChefs: number;
  ordersByStatus: AdminStatsOrdersByStatus;
  revenueByDay: AdminStatsRevenueByDayItem[];
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  placed: "placed",
  preparing: "preparing",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export interface UpsertChefRequest {
  name: string;
  tagline: string;
  cuisine: string;
  rating: number;
  etaMinutes: number;
  deliveryFee: number;
  imageUrl: string;
  location: string;
  priceForTwo: number;
  isVeg: boolean;
  featured?: boolean;
}

export interface UpsertProductRequest {
  name: string;
  description: string;
  price: number;
  mrp: number;
  unit: string;
  imageUrl: string;
  categoryId: number;
  inStock?: boolean;
  essential?: boolean;
}

export type CreateOrderRequestPaymentMethod =
  (typeof CreateOrderRequestPaymentMethod)[keyof typeof CreateOrderRequestPaymentMethod];

export const CreateOrderRequestPaymentMethod = {
  cod: "cod",
  online: "online",
} as const;

export interface CreateOrderRequest {
  deviceId: string;
  items: OrderItem[];
  address: Address;
  paymentMethod: CreateOrderRequestPaymentMethod;
  paymentReference?: string;
  deliveryFee: number;
  tip?: number;
  notes?: string;
}

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  cod: "cod",
  online: "online",
} as const;

export type OrderPaymentStatus =
  (typeof OrderPaymentStatus)[keyof typeof OrderPaymentStatus];

export const OrderPaymentStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  placed: "placed",
  preparing: "preparing",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  deviceId: string;
  items: OrderItem[];
  address: Address;
  paymentMethod: OrderPaymentMethod;
  paymentStatus: OrderPaymentStatus;
  paymentReference?: string;
  status: OrderStatus;
  subtotal: number;
  deliveryFee: number;
  tip: number;
  total: number;
  notes?: string;
  createdAt: string;
}

export interface ProcessPaymentRequest {
  amount: number;
  /** Simulated card number; ends with even digit = success, odd = fail */
  cardNumber: string;
  cardHolder: string;
  expiry: string;
  cvv: string;
}

export interface PaymentResult {
  success: boolean;
  reference: string;
  message: string;
}

export type ListChefsParams = {
  cuisine?: string;
  q?: string;
};

export type ListProductsParams = {
  categoryId?: number;
  q?: string;
};

export type SearchAllParams = {
  q: string;
};

export type AdminListOrdersParams = {
  status?: string;
};

export type ListOrdersParams = {
  deviceId: string;
};