# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HomeBites - Home food and groceries marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chefs:
    get:
      operationId: listChefs
      tags: [catalog]
      summary: List home chefs
      parameters:
        - in: query
          name: cuisine
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: Chefs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chef"
  /chefs/{id}:
    get:
      operationId: getChef
      tags: [catalog]
      summary: Get a home chef with their menu
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chef detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChefDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: List grocery products
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [catalog]
      summary: Get a grocery product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /search:
    get:
      operationId: searchAll
      tags: [catalog]
      summary: Search across home food dishes, chefs, and groceries
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Admin dashboard stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/orders/{id}/status:
    patch:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Update an order's status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/chefs:
    get:
      operationId: adminListChefs
      tags: [admin]
      summary: Admin list chefs
      responses:
        "200":
          description: Chefs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chef"
    post:
      operationId: adminCreateChef
      tags: [admin]
      summary: Create a chef
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertChefRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chef"
  /admin/chefs/{id}:
    patch:
      operationId: adminUpdateChef
      tags: [admin]
      summary: Update a chef
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertChefRequest"
      responses:
        "200":
          description: Updated chef
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chef"
    delete:
      operationId: adminDeleteChef
      tags: [admin]
      summary: Delete a chef
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: Admin list products
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProductRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Chef:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        tagline:
          type: string
        cuisine:
          type: string
        rating:
          type: number
        etaMinutes:
          type: integer
        deliveryFee:
          type: number
        imageUrl:
          type: string
        location:
          type: string
        priceForTwo:
          type: number
        isVeg:
          type: boolean
        featured:
          type: boolean
      required:
        [
          id,
          name,
          tagline,
          cuisine,
          rating,
          etaMinutes,
          deliveryFee,
          imageUrl,
          location,
          priceForTwo,
          isVeg,
        ]
    Dish:
      type: object
      properties:
        id:
          type: integer
        chefId:
          type: integer
        chefName:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        isVeg:
          type: boolean
        rating:
          type: number
        spiceLevel:
          type: string
          enum: [mild, medium, spicy]
      required:
        [id, chefId, chefName, name, description, price, imageUrl, isVeg, rating, spiceLevel]
    ChefDetail:
      type: object
      properties:
        chef:
          $ref: "#/components/schemas/Chef"
        dishes:
          type: array
          items:
            $ref: "#/components/schemas/Dish"
      required: [chef, dishes]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        mrp:
          type: number
        unit:
          type: string
        imageUrl:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        inStock:
          type: boolean
        essential:
          type: boolean
      required:
        [
          id,
          name,
          description,
          price,
          mrp,
          unit,
          imageUrl,
          categoryId,
          categoryName,
          inStock,
        ]
...
    SearchResults:
      type: object
      properties:
        chefs:
          type: array
          items:
            $ref: "#/components/schemas/Chef"
        dishes:
          type: array
          items:
            $ref: "#/components/schemas/Dish"
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required: [chefs, dishes, products]
    OrderItem:
      type: object
      properties:
        kind:
          type: string
          enum: [dish, product]
        refId:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
        unitPrice:
          type: number
        quantity:
          type: integer
      required: [kind, refId, name, imageUrl, unitPrice, quantity]
    Address:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        pincode:
          type: string
        landmark:
          type: string
      required: [fullName, phone, line1, city, pincode]
...
    AdminStats:
      type: object
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        totalProducts:
          type: integer
        totalChefs:
          type: integer
        ordersByStatus:
          type: object
          additionalProperties:
            type: integer
        revenueByDay:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
              revenue:
                type: number
              orders:
                type: integer
            required: [day, revenue, orders]
      required:
        [
          totalOrders,
          totalRevenue,
          totalProducts,
          totalChefs,
          ordersByStatus,
          revenueByDay,
        ]
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [placed, preparing, out_for_delivery, delivered, cancelled]
      required: [status]
    UpsertChefRequest:
      type: object
      properties:
        name:
          type: string
        tagline:
          type: string
        cuisine:
          type: string
        rating:
          type: number
        etaMinutes:
          type: integer
        deliveryFee:
          type: number
        imageUrl:
          type: string
        location:
          type: string
        priceForTwo:
          type: number
        isVeg:
          type: boolean
        featured:
          type: boolean
      required:
        [
          name,
          tagline,
          cuisine,
          rating,
          etaMinutes,
          deliveryFee,
          imageUrl,
          location,
          priceForTwo,
          isVeg,
        ]
    UpsertProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        mrp:
          type: number
        unit:
          type: string
        imageUrl:
          type: string
        categoryId:
          type: integer
        inStock:
          type: boolean
        essential:
          type: boolean
      required:
        [name, description, price, mrp, unit, imageUrl, categoryId]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        address:
          $ref: "#/components/schemas/Address"
        paymentMethod:
          type: string
          enum: [cod, online]
        paymentStatus:
          type: string
          enum: [pending, paid, failed]
        paymentReference:
          type: string
        status:
          type: string
          enum: [placed, preparing, out_for_delivery, delivered, cancelled]
        subtotal:
          type: number
        deliveryFee:
          type: number
        tip:
          type: number
        total:
          type: number
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        [
          id,
          deviceId,
          items,
          address,
          paymentMethod,
          paymentStatus,
          status,
          subtotal,
          deliveryFee,
          tip,
          total,
          createdAt,
        ]
...