# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HomeBites - Home food and groceries marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsOrdersByStatus } from "./adminStatsOrdersByStatus";
import type { AdminStatsRevenueByDayItem } from "./adminStatsRevenueByDayItem";

export interface AdminStats {
  totalOrders: number;
  totalRevenue: number;
  totalProducts: number;
  totalChefs: number;
  ordersByStatus: AdminStatsOrdersByStatus;
  revenueByDay: AdminStatsRevenueByDayItem[];
}